Fix reading recreated measurements file
When we are reading that file too fast MSI Afterburner stops writing to it. This happens when log file is re-created and a lot of changes are made in a very short time because of the writing of the header. Old measurements reading implementation was not ready for this scenario and made to many open/read requests for the log file, most probably  blocking Afterburner from opening/writing to the log file.

This commit fixed this issue by dropping concurrent read requests and introducing delay (100ms) before each read and longer delay after reading exceptions (1s). This eliminates the problem. User experience was not downgraded because introduced delay is too small to be perceived and readings are always fresh.